// src/app/layout/trcm-lab-automation/models/shipped/shipped.ts
export class Shipped {

    action: string;
    actionStatus: string;
    board_defect: string;
    bt_address: string;
    creationOn: Date;
    devices: {
      action: '';
      actionStatus: '';
      contract_expires: Date;
      contract_starts: Date;
      contractperiod: number;
      creationDate: '';
      deviceContractStatus: '';
      deviceModels: '';
      deviceWarantyStatus: '';
      deviceowner: '';
      id: number;
      imeinumber: '';
      intrash: '';
      manufacturer: '';
      partnumber: '';
      seller: '';
      serialnumber: '';
      warrantyExpire: Date;
      warranty_starts: Date;
      warrantyperiod: number;
    };
    id: number;
    intrash: string;
    mac_id: string;
    note: string;
    pcba_pn: string;
    pcba_sn: string;
    pki_version: string;
    printer_type: string;
    serialNumber: string;
    shippedStatus: string;
    wifi: string;

  }